requirements.txt
ccxt==4.2.25
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0